build: switch js & quality checks off of paver
