+++ .github/workflows/js-tests.yml
@@ -64,13 +64,15 @@ jobs:
           make base-requirements
 
       - uses: c-hive/gha-npm-cache@v1
+
+      - name: Install npm
+        run: npm ci  
+
       - name: Run JS Tests
-        env:
-          TEST_SUITE: js-unit
-          SCRIPT_TO_RUN: ./scripts/generic-ci-tests.sh
         run: |
           npm install -g jest
-          xvfb-run --auto-servernum ./scripts/all-tests.sh
+          xvfb-run --auto-servernum make test-js
+          make coverage-js
 
       - name: Save Job Artifacts
         uses: actions/upload-artifact@v4
