+++ .github/workflows/quality-checks.yml
@@ -60,16 +60,30 @@ jobs:
           PIP_SRC: ${{ runner.temp }}
         run: |
           make test-requirements
-
+      
+      - name: Install npm
+        env:
+          PIP_SRC: ${{ runner.temp }}
+        run: npm ci  
+      
+      - name: Install python packages
+        env:
+          PIP_SRC: ${{ runner.temp }}
+        run: |
+            pip install -e .
+      
       - name: Run Quality Tests
         env:
-          TEST_SUITE: quality
-          SCRIPT_TO_RUN: ./scripts/generic-ci-tests.sh
           PIP_SRC: ${{ runner.temp }}
           TARGET_BRANCH: ${{ github.base_ref }}
         run: |
-          ./scripts/all-tests.sh
-
+            make pycodestyle
+            make eslint
+            make stylelint
+            make xsslint
+            make pii_check
+            make check_keywords
+          
       - name: Save Job Artifacts
         if: always()
         uses: actions/upload-artifact@v4
