--- pavelib/utils/test/suites/__init__.py
@@ -1,5 +1,0 @@
-"""
-TestSuite class and subclasses
-"""
-from .js_suite import JestSnapshotTestSuite, JsTestSuite
-from .suite import TestSuite
